refactor: use a universal component to generate md like post content.
- take showComments and showThumbnail prompts to optionally show thumbnail or comments.
